Updated view classes, modified model classes and controller.
Additional changes to begin linking model classes to the view.  Created
game menu view, added functionality so that the user starts at the
game_menu, makes selections, and is redirect to the game_main screen.
Changes were necessary in several of the model classes and the
controller to make this new functionality work.